Données de Firebase récupérées + la bonne loc s'affiche sur la map (plus de pb de CGFloat) mais seulement depuis le side menu

+++ fonrose-ecommerceV2/Views/Location/LocationViewModel.swift
@@ -216,7 +216,7 @@ class LocationViewModel: NSObject, ObservableObject, CLLocationManagerDelegate {
         
     }
     
-    func pinSelectedPlace(pointSelectedPlaceLat: Int, pointSelectedPlaceLong: Int) {
+    func pinSelectedPlace(pointSelectedPlaceLat: CGFloat, pointSelectedPlaceLong: CGFloat) {
                   
         //la variable "coordinate" correspond aux coordonnées d'un objet place de type "Place"
         let selectedPlaceCoordinate = CLLocationCoordinate2D(latitude: CLLocationDegrees(pointSelectedPlaceLat), longitude: CLLocationDegrees(pointSelectedPlaceLong))
